+++ app/apirequests/contentCard-apirequests-request.php
@@ -1,0 +1,37 @@
+<?php
+// Card
+$cardHeaderContent = $mmAPIRequestsT->FontAwesome . STR_NBSP . $mmAPIRequestsT->NameSingular;
+$card = new \xan\eleCard( CARD_WIDTH_2X, '', false );
+
+// Tags Special
+$tagsEleInput_Notes = new \xan\tags( [ 'col', ], [ 'height' => ELE_HEIGHT_6X ], [] );
+
+// Table
+$table = new \xan\eleTable( $tagsCellEmpty );
+$tableRowIndex = -1;
+
+// Table Rows
+$table->cellSet( ++$tableRowIndex, 0, $tagsCellRightMiddle, $mmAPIRequestsT->getColEleRender( 'Auth', ELE_AS_LABEL, $tagsEleLabel, $recsDetail, $formTagDetail, $resp ) );
+$table->cellSet( $tableRowIndex, 1, $tagsCellLeftMiddle, $mmAPIRequestsT->getColEleRender( 'Auth', ELE_AS_DEFINED, $tagsEleInput, $recsDetail, $formTagDetail, $resp ) );
+
+$table->cellSet( ++$tableRowIndex, 0, $tagsCellRightMiddle, $mmAPIRequestsT->getColEleRender( 'Action', ELE_AS_LABEL, $tagsEleLabel, $recsDetail, $formTagDetail, $resp ) );
+$table->cellSet( $tableRowIndex, 1, $tagsCellLeftMiddle, $mmAPIRequestsT->getColEleRender( 'Action', ELE_AS_DEFINED, $tagsEleInput, $recsDetail, $formTagDetail, $resp ) );
+
+$table->cellSet( ++$tableRowIndex, 0, $tagsCellRightMiddle, $mmAPIRequestsT->getColEleRender( 'ActionID', ELE_AS_LABEL, $tagsEleLabel, $recsDetail, $formTagDetail, $resp ) );
+$table->cellSet( $tableRowIndex, 1, $tagsCellLeftMiddle, $mmAPIRequestsT->getColEleRender( 'ActionID', ELE_AS_DEFINED, $tagsEleInput, $recsDetail, $formTagDetail, $resp ) );
+
+$table->cellSet( ++$tableRowIndex, 0, $tagsCellRightMiddle, $mmAPIRequestsT->getColEleRender( 'RequestTS', ELE_AS_LABEL, $tagsEleLabel, $recsDetail, $formTagDetail, $resp ) );
+$table->cellSet( $tableRowIndex, 1, $tagsCellLeftMiddle, $mmAPIRequestsT->getColEleRender( 'RequestTS', ELE_AS_DEFINED, $tagsEleInput, $recsDetail, $formTagDetail, $resp ) );
+
+$table->cellSet( ++$tableRowIndex, 0, $tagsCellRightMiddle, $mmAPIRequestsT->getColEleRender( 'RequestIsProcessed', ELE_AS_LABEL, $tagsEleLabel, $recsDetail, $formTagDetail, $resp ) );
+$table->cellSet( $tableRowIndex, 1, $tagsCellLeftMiddle, $mmAPIRequestsT->getColEleRender( 'RequestIsProcessed', ELE_AS_DEFINED, $tagsEleInput, $recsDetail, $formTagDetail, $resp ) );
+
+$table->cellSet( ++$tableRowIndex, 0, $tagsCellRightMiddle, $mmAPIRequestsT->getColEleRender( 'RequestIsSent', ELE_AS_LABEL, $tagsEleLabel, $recsDetail, $formTagDetail, $resp ) );
+$table->cellSet( $tableRowIndex, 1, $tagsCellLeftMiddle, $mmAPIRequestsT->getColEleRender( 'RequestIsSent', ELE_AS_DEFINED, $tagsEleInput, $recsDetail, $formTagDetail, $resp ) );
+
+$table->cellSet( ++$tableRowIndex, 0, $tagsCellRightMiddle, $mmAPIRequestsT->getColEleRender( 'RequestData', ELE_AS_LABEL, $tagsEleLabel, $recsDetail, $formTagDetail, $resp ) );
+$table->cellSet( $tableRowIndex, 1, $tagsCellLeftMiddle, $mmAPIRequestsT->getColEleRender( 'RequestData', ELE_AS_DEFINED, $tagsEleInput_Notes, $recsDetail, $formTagDetail, $resp ) );
+
+// Card Append
+$resp->contentAreaA[] = $card->renderCardWithDiv( $cardHeaderContent, $table->render() );
+?>