+++ app/zzMetaModules/mmAPIRequestsT.php
@@ -11,12 +11,12 @@ class moduleMetaAPIRequestsT extends \xan\moduleMeta {
 		$this->NameTableParam = \xan\strSubstitute( $this->NameTableKey, 'UU', '' );
 		
 		// QueryBuilder Operators: equal, not_equal, in, not_in, less, less_or_equal, greater, greater_or_equal, between, not_between, begins_with, not_begins_with, contains, not_contains, ends_with, not_ends_with, is_empty, is_not_empty, is_null, is_not_null
-		// $this->QuerySimpleDefault = array( 'NameCompany', 'NameLast', 'NameFirst', 'EmailAddress', 'PhoneWork', 'PhoneMobile' );
-		// $this->QueryBuilderDefault = '{ field: "Users.NameFirst", id: "querybuilder_Users_NameFirst", operator: "begins_with", value: "" }';
-		//
-		// $this->QueryOrderByDefault = 'Active DESC, PrivAdmin DESC, NameLast ASC, NameFirst ASC';
-		// $this->QueryOrderByExtraBegin .= \xan\dbQueryOrderByItem( $this->NameTable, 'Active DESC, PrivAdmin DESC, NameLast ASC, NameFirst ASC', 'Active ASC, PrivAdmin ASC, NameLast ASC, NameFirst ASC', 'Admin, Active' );
-		// $this->QueryOrderByExtraBegin .= '<hr />';
+		$this->QuerySimpleDefault = array( 'Auth', 'Action' );
+		$this->QueryBuilderDefault = '{ field: "APIRequests.RequestTS", id: "querybuilder_APIRequests_RequestTS", operator: "greater_or_equal", value: "" }';
+		
+		$this->QueryOrderByDefault = 'RequestTS DESC';
+		$this->QueryOrderByExtraBegin .= \xan\dbQueryOrderByItem( $this->NameTable, 'RequestTS DESC', 'RequestTS ASC', 'Request TS' );
+		$this->QueryOrderByExtraBegin .= '<hr />';
 		
 		$this->NamePlural = 'API Requests';
 		$this->NameSingular = 'API Request';
@@ -35,21 +35,112 @@ class moduleMetaAPIRequestsT extends \xan\moduleMeta {
 	///////////////////////////////////////////////////////////
 	// Functions Required by \xan\moduleMeta
 	
-	public function getDisplayList( \xan\recs $recs ) {
-		$code = $this->getDisplayName( $recs );
+	public function getDisplayName( \xan\recs $recs ) {
+		$code = trim( $recs->rowsD[ $recs->rowIndex ][ 'RequestTS' ] . " " . $recs->rowsD[ $recs->rowIndex ][ 'Auth' ] . " " . $recs->rowsD[ $recs->rowIndex ][ 'Action' ] . " " . $recs->rowsD[ $recs->rowIndex ][ 'ActionID' ] );
+		
 		$code = trim( $code );
 		return $code;
 	}
 	
-	public function getDisplayName( \xan\recs $recs ) {
-		$code = '';
+	public function getDisplayList( \xan\recs $recs ) {
+		$code = $this->getDisplayName( $recs );
+		
 		$code = trim( $code );
 		return $code;
 	}
 	
 	public function getListItem( $idPrefix, \xan\recs $recs, $onClick ) {
-		$code = '';
+		$idListItem = $idPrefix . $recs->rowsD[ $recs->rowIndex ][ $this->NameTableKey ];
+		$idListItemImage = $idListItem . 'Image';
+		$idListItemLabel = $idListItem . 'Label';
+		
+		// Table Init
+		$tagsCellEmpty = new \xan\tags( [ 'border-0', 'pb-0', TEXT_ALIGN_LEFT, TABLE_ALIGN_MIDDLE ], [], [] );
+		$tagsCellRightMiddle = new \xan\tags( [ 'border-0', 'pb-0', TEXT_ALIGN_LEFT, TABLE_ALIGN_TOP ], [], [] );
+		$table = new \xan\eleTable( $tagsCellEmpty );
+		
+		// Table Cell
+		$info = '<span id="' . $idListItemLabel . '" class="list-group-item-text">' . $this->getDisplayList( $recs ) . '</span>';
+		$table->cellSet( $recs->rowIndex, 0, $tagsCellRightMiddle, $info );
+		
+		// Content
+		$code = $table->render();
 		return $code;
+	}
+	
+	public function getListItemRow( $idPrefix, \xan\recs $recs, $onClick, \xan\eleTable &$table, $idSselected ) {
+		$idListItem = $idPrefix . $recs->rowsD[ $recs->rowIndex ][ $this->NameTableKey ];
+		$isActive = ( $recs->rowsD[ $recs->rowIndex ][ $this->NameTableKey ] === $idSselected ? 'active' : '' );
+		$colIndex = 0;
+		$rowIndexTable = $recs->rowIndex + 1;
+		
+		// Cell Tag
+		$tagsCell = new \xan\tags( [ $isActive, 'border-0', 'pb-0', TEXT_ALIGN_LEFT, TABLE_ALIGN_TOP ], [], [] );
+		
+		// Cell Left
+		$buttonMoreTags = new \xan\tags( [ ELE_CLASS_BUTTON_SM_SECONDARY, 'mb-2' ], [], [ 'id="' . $idListItem . '"', 'onclick="window.location.href = \'' . $this->URLFull . $recs->rowsD[ $recs->rowIndex ][ $this->NameTableKey ] . '\';"' ] );
+		$buttonMoreEle = new \xan\eleButton( $rowIndexTable, '', '', $buttonMoreTags );
+		$table->cellSet( $rowIndexTable, $colIndex, $tagsCell, '<div style="height: 100px; overflow-y: auto;">' . $buttonMoreEle->render() . '</div>' );
+		
+		// Cells
+		foreach ( $recs->rowsD[ $recs->rowIndex ] as $key => $value ) {
+			$colIndex++;
+			switch ( $key ) {
+				case 'RequestData':
+					$width = 400;
+					break;
+				case 'ResponseURL':
+					$width = 200;
+					break;
+				case 'ResponseData':
+					$width = 500;
+					break;
+				case 'Log':
+					$width = 300;
+					break;
+				default:
+					$width = 100;
+			}
+			
+			// $tagsCell = new \xan\tags( [ $isActive, 'border-0', 'pb-0', TEXT_ALIGN_LEFT, TABLE_ALIGN_TOP ], [], [] );
+			$table->cellSet( $rowIndexTable, $colIndex, $tagsCell, '<div style="height: 100px; width: ' . $width . 'px; overflow-y: auto;">' . $value . '</div>' );
+		}
+		
+	}
+	
+	public function getListItemRowHeader( $idPrefix, \xan\recs $recs, $onClick, \xan\eleTable &$table, $idSselected ) {
+		$colIndex = 0;
+		
+		// Cell Tag
+		$tagsCell = new \xan\tags( [ 'border-0', 'pb-0', TEXT_ALIGN_LEFT, TABLE_ALIGN_TOP ], [ 'position' => '-webkit-sticky', 'position' => '-moz-sticky', 'position' => '-ms-sticky', 'position' => '-o-sticky', 'position' => 'sticky', 'top' => 0 ], [] );
+		
+		// Cell Left
+		$table->cellSet( $recs->rowIndex, $colIndex, $tagsCell, '<div style="overflow-y: auto; "></div>' );
+		
+		// Cells
+		foreach ( $recs->rowsD[ $recs->rowIndex ] as $key => $value ) {
+			$colIndex++;
+			switch ( $key ) {
+				case 'RequestData':
+					$width = 400;
+					break;
+				case 'ResponseURL':
+					$width = 200;
+					break;
+				case 'ResponseData':
+					$width = 500;
+					break;
+				case 'Log':
+					$width = 300;
+					break;
+				default:
+					$width = 100;
+			}
+			
+			// $tagsCell = new \xan\tags( [ 'border-0', 'pb-0', TEXT_ALIGN_LEFT, TABLE_ALIGN_TOP ], [], [] );
+			$table->cellSet( $recs->rowIndex, $colIndex, $tagsCell, '<div style="width: ' . $width . 'px; overflow-y: auto;">' . $key . '</div>' );
+		}
+		
 	}
 	
 	public function getColLabel( $colName ) {
@@ -76,15 +167,6 @@ class moduleMetaAPIRequestsT extends \xan\moduleMeta {
 		
 		// Columns Specifics
 		switch ( $colName ) {
-			case 'Active':
-				$colMeta->colLabelEN = 'Active';
-				$colMeta->eleType = ELE_TYPE_SELECT_DB;
-				$colMeta->choicesAValues = ARRAY_YESNO;
-				$colMeta->choicesADisplay = ARRAY_YESNO;
-				$colMeta->choicesOtherLabel = '';
-				break;
-			
-			// App
 			case 'Auth':
 				$colMeta->colLabelEN = 'Request Auth';
 				break;
@@ -114,6 +196,7 @@ class moduleMetaAPIRequestsT extends \xan\moduleMeta {
 				break;
 			case 'RequestData':
 				$colMeta->colLabelEN = 'Request Data';
+				$colMeta->eleType = ELE_TYPE_TEXTAREA_DB;
 				break;
 			case 'ResponseTS':
 				$colMeta->colLabelEN = 'Response TS';
@@ -121,6 +204,7 @@ class moduleMetaAPIRequestsT extends \xan\moduleMeta {
 				break;
 			case 'ResponseData':
 				$colMeta->colLabelEN = 'Response Data';
+				$colMeta->eleType = ELE_TYPE_TEXTAREA_DB;
 				break;
 			case 'ResponseURL':
 				$colMeta->colLabelEN = 'Response URL';